feat(step): add new supply step
