+++ step-04/config.go
@@ -1,0 +1,5 @@
+package main
+
+type Config struct {
+	Port int
+}
