+++ step-04/go.mod
@@ -2,10 +2,7 @@ module dojo-fx
 
 go 1.21.3
 
-require (
-	github.com/go-chi/chi/v5 v5.0.12
-	go.uber.org/fx v1.21.0
-)
+require go.uber.org/fx v1.21.0
 
 require (
 	go.uber.org/dig v1.17.1 // indirect
