+++ step-04/main.go
@@ -7,6 +7,7 @@ import (
 func main() {
 	fx.New(
 		fx.Provide(NewServer),
-		fx.Invoke(StartServer),
+		fx.Invoke(ServerStart),
+		fx.Supply(&Config{Port: 8080}),
 	).Run()
 }
