--- step-04/router.go
@@ -1,9 +1,0 @@
-package main
-
-import "github.com/go-chi/chi/v5"
-
-func NewRouter() *chi.Mux {
-	router := chi.NewRouter()
-
-	return router
-}
