+++ step-04/server.go
@@ -1,30 +1,14 @@
 package main
 
 import (
-	"context"
-	"errors"
-	"go.uber.org/fx"
+	"fmt"
 	"net/http"
 )
 
-func NewServer() *http.Server {
-	return &http.Server{Addr: ":8080"}
+func NewServer(cfg *Config) *http.Server {
+	return &http.Server{Addr: fmt.Sprintf(":%d", cfg.Port)}
 }
 
-func StartServer(lc fx.Lifecycle, server *http.Server) {
-	lc.Append(fx.Hook{
-		OnStart: func(_ context.Context) error {
-			go func() {
-				if err := server.ListenAndServe(); err != nil {
-					if !errors.Is(err, http.ErrServerClosed) {
-						panic(err)
-					}
-				}
-			}()
-			return nil
-		},
-		OnStop: func(ctx context.Context) error {
-			return server.Shutdown(ctx)
-		},
-	})
+func ServerStart(server *http.Server) {
+	fmt.Printf("Server started on %s\n", server.Addr)
 }
