+++ step-05/README.md
@@ -1,30 +1,37 @@
-# Step 05: Add a logger
+# Step 04: Add a router and a first handler
 
-Now we want to be able to log information to know what's happening in our application.
+We now want to be able to handle a request on a `/status` route and respond with the `OK` string.
 
 ## Documentation
 
-In the `logger.go` file, you'll find a new function to create a new logger.
+In the `router.go` file, you'll find a new function to create a new Chi router.
 
-This logger can be used as follows:
+This router can be given to our server as the main router to handle requests:
 
 ```go
-logger.Info().Msg("An info message")
-logger.Warn().Msg("A warning message")
-logger.Error().Msg("An error message")
+http.Server{Addr: ":8080", Handler: chiRouter}
+```
+
+You can provide a handler for a specific route using the `Method` function on the router:
+
+```go
+chiRouter.Method(http.MethodGet, "/status", statusHandler)
 ```
 
 ## Tasks
 
-1. Use the `NewLogger` function to provide a logger to the app
-2. Use the logger to log information about the server starting and stopping
-3. Use the logger to log information about the server receiving a status request
+1. Provide the `NewRouter` function in the FX app
+2. Use the router in the `NewServer` function as an HTTP server handler
+3. Create a new `StatusHandler` struct implementing the `http.Handler` interface, it should respond with the `OK` string
+4. Create a `NewStatusHandler` function and provide it to the FX app
+5. Update the `NewRouter` function to use the `StatusHandler` for the `GET /status` route
 
-Now run your application and call the `/status` endpoint to see if the logger is working.
+Run the application, the following command should return `OK`:
 
-Our goal is to see something like this:
+```sh
+curl http://localhost:8080/status
+```
 
-![Logger output](./assets/logs.png)
+Any other route should return a `404 Not Found` error.
 
-
-If you get the right output, you shall now proceed to the [next step](../step-06/README.md).
+If you get the right responses, you shall now proceed to the [next step](../step-06/README.md).
