+++ step-05/main.go
@@ -8,7 +8,6 @@ func main() {
 	fx.New(
 		fx.Provide(NewServer),
 		fx.Invoke(StartServer),
-		fx.Provide(NewRouter),
-		fx.Provide(NewStatusHandler),
+		fx.Supply(&Config{Port: 8080}),
 	).Run()
 }
