+++ step-05/router.go
@@ -2,10 +2,8 @@ package main
 
 import "github.com/go-chi/chi/v5"
 
-func NewRouter(sh *StatusHandler) *chi.Mux {
+func NewRouter() *chi.Mux {
 	router := chi.NewRouter()
-
-	router.Method("GET", "/status", sh)
 
 	return router
 }
