+++ step-05/status_handler.go
@@ -1,14 +1,3 @@
 package main
 
-import "net/http"
-
-type StatusHandler struct {
-}
-
-func NewStatusHandler() *StatusHandler {
-	return &StatusHandler{}
-}
-
-func (h *StatusHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	w.Write([]byte("OK"))
-}
+// Hum, something is missing here...
