+++ step-06/main.go
@@ -8,8 +8,8 @@ func main() {
 	fx.New(
 		fx.Provide(NewServer),
 		fx.Invoke(StartServer),
+		fx.Supply(&Config{Port: 8080}),
 		fx.Provide(NewRouter),
 		fx.Provide(NewStatusHandler),
-		fx.Provide(NewLogger),
 	).Run()
 }
