+++ step-07/main.go
@@ -8,16 +8,9 @@ func main() {
 	fx.New(
 		fx.Provide(NewServer),
 		fx.Invoke(StartServer),
+		fx.Supply(&Config{Port: 8080}),
 		fx.Provide(NewRouter),
-		fx.Provide(AsRouteHandler(NewStatusHandler)),
+		fx.Provide(NewStatusHandler),
 		fx.Provide(NewLogger),
 	).Run()
-}
-
-func AsRouteHandler(h any) any {
-	return fx.Annotate(
-		h,
-		fx.As(new(RouteHandler)),
-		// Hum, something is missing here...
-	)
 }
