+++ step-07/route_handler.go
@@ -1,8 +1,6 @@
 package main
 
-import (
-	"net/http"
-)
+import "net/http"
 
 type Route struct {
 	Method string
