+++ step-08/main.go
@@ -7,11 +7,11 @@ import (
 func main() {
 	fx.New(
 		fx.Provide(NewServer),
-		fx.Provide(AsRouteHandler(NewHelloHandler)),
-		fx.Provide(AsRouteHandler(NewStatusHandler)),
-		fx.Provide(fx.Annotate(NewRouter, fx.ParamTags(`group:"routeHandlers"`))),
-		fx.Provide(NewLogger),
 		fx.Invoke(StartServer),
+		fx.Supply(&Config{Port: 8080}),
+		fx.Provide(NewRouter),
+		fx.Provide(AsRouteHandler(NewStatusHandler)),
+		fx.Provide(NewLogger),
 	).Run()
 }
 
@@ -19,6 +19,6 @@ func AsRouteHandler(h any) any {
 	return fx.Annotate(
 		h,
 		fx.As(new(RouteHandler)),
-		fx.ResultTags(`group:"routeHandlers"`),
+		// Hum, something is missing here...
 	)
 }
