+++ step-09/status_handler.go
@@ -9,13 +9,22 @@ type StatusHandler struct {
 	logger *zerolog.Logger
 }
 
-func NewStatusHandler(logger *zerolog.Logger) *StatusHandler {
-	return &StatusHandler{
-		logger: logger,
+func (h *StatusHandler) Pattern() []Route {
+	return []Route{
+		{
+			Path:   "/status",
+			Method: http.MethodGet,
+		},
 	}
 }
 
 func (h *StatusHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	h.logger.Info().Msg("StatusHandler called")
 	w.Write([]byte("OK"))
+}
+
+func NewStatusHandler(logger *zerolog.Logger) *StatusHandler {
+	return &StatusHandler{
+		logger: logger,
+	}
 }
