+++ step-10/README.md
@@ -1,0 +1,7 @@
+# This is the end
+
+This final step provide the code solution to the previous one.
+
+Congratulations on getting this far!
+
+I hope you enjoyed this journey and learned something about Go and FX.
