+++ step-10/config/config.go
@@ -1,0 +1,5 @@
+package config
+
+type Config struct {
+	Port int
+}
