+++ step-10/config/module.go
@@ -1,0 +1,10 @@
+package config
+
+import "go.uber.org/fx"
+
+var Module = fx.Module(
+	"http handlers",
+	fx.Supply(&Config{
+		Port: 8080,
+	}),
+)
