+++ step-10/go.mod
@@ -1,13 +1,16 @@
-module dojo-fx
+module dojo-fx/step-10
 
 go 1.21.3
 
 require (
 	github.com/go-chi/chi/v5 v5.0.12
+	github.com/rs/zerolog v1.32.0
 	go.uber.org/fx v1.21.0
 )
 
 require (
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.19 // indirect
 	go.uber.org/dig v1.17.1 // indirect
 	go.uber.org/multierr v1.11.0 // indirect
 	go.uber.org/zap v1.27.0 // indirect
