+++ step-10/handler/hello.go
@@ -1,7 +1,7 @@
 package handler
 
 import (
-	"dojo-fx/step-09/repository"
+	"dojo-fx/step-10/repository"
 	"github.com/rs/zerolog"
 	"net/http"
 )
