+++ step-10/http/server.go
@@ -2,6 +2,7 @@ package http
 
 import (
 	"context"
+	"dojo-fx/step-10/config"
 	"errors"
 	"fmt"
 	"github.com/go-chi/chi/v5"
@@ -10,8 +11,8 @@ import (
 	"net/http"
 )
 
-func NewServer(router *chi.Mux) *http.Server {
-	return &http.Server{Addr: ":8080", Handler: router}
+func NewServer(router *chi.Mux, cfg *config.Config) *http.Server {
+	return &http.Server{Addr: fmt.Sprintf(":%d", cfg.Port), Handler: router}
 }
 
 func StartServer(lc fx.Lifecycle, server *http.Server, logger *zerolog.Logger, shutdowner fx.Shutdowner) {
